chore(Features): move type_conversion to submodule
