+++ src/config/deserialization/config_options/features.rs
@@ -1,7 +1,8 @@
 use crate::Result;
-use os_checker_types::config as out;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+
+mod type_conversion;
 
 #[derive(Debug, Serialize, Deserialize, JsonSchema, Clone)]
 #[serde(untagged)]
@@ -92,57 +93,8 @@ where
     let features = <&str>::deserialize(deserializer)?;
     Ok(features.split(',').map(String::from).collect())
 }
-
-// ******************** type conversion ********************
-
-impl From<FeaturesCompleteState> for out::FeaturesCompleteState {
-    fn from(
-        FeaturesCompleteState {
-            f,
-            no_default_features,
-            all_features,
-            targets,
-        }: FeaturesCompleteState,
-    ) -> Self {
-        Self {
-            f: f.into(),
-            no_default_features,
-            all_features,
-            targets,
-        }
-    }
 }
 
-impl From<FeaturesWithCommas> for out::FeaturesWithCommas {
-    fn from(value: FeaturesWithCommas) -> Self {
-        Self {
-            features: value.features,
-        }
-    }
 }
 
-impl From<out::FeaturesCompleteState> for FeaturesCompleteState {
-    fn from(
-        out::FeaturesCompleteState {
-            f,
-            no_default_features,
-            all_features,
-            targets,
-        }: out::FeaturesCompleteState,
-    ) -> Self {
-        Self {
-            f: f.into(),
-            no_default_features,
-            all_features,
-            targets,
-        }
-    }
-}
-
-impl From<out::FeaturesWithCommas> for FeaturesWithCommas {
-    fn from(value: out::FeaturesWithCommas) -> Self {
-        Self {
-            features: value.features,
-        }
-    }
 }
